feat: main config forwards to env config
